Fixed potential crash when triggering AutoMap
Keeping around the AutoMapper instances was introduced as an
optimization, based on the assumption that they only kept the
automapping rules they would apply.

In fact, because they also store a pointer to the map to which the rules
are to be applied, triggering AutoMap for a second map could lead to
Tiled crashing.

Ideally, the AutoMapper class would be split up in a part that can be
kept around (storing the rule maps) and a part that is re-initialized
each time AutoMapping is applied. This way we would avoid needless
reloading of the rule maps when the same rules are applied to a
different map.

Closes #2766