Introduced ProjectManager singleton
The ScriptManager depends on the project to get the project-specific
extensions folder. However, the project was owned by the ProjectModel,
which was owned by the ProjectView, which was owned by the ProjectDock,
which was owned by the MainWindow. And the MainWindow is not created
when scripts are used for exporting on the command-line. Now the
ScriptManager can instead talk to the ProjectManager directly.

Similarly, the ScriptModule depended on the DocumentManager, which was
since recently however only instantiated as part of the MainWindow.
Since the relevant part of the scripting API is not useful when running
export scripts on the command-line, this dependency was made optional.

This change also reduces the rather hacky direct use of the MainWindow.

Closes #2842