Scripting: Fixed issues with absolute file paths on Windows
The strings were converted into URLs was broken on Windows, where an
absolute file path like "C:/..." was interpreted as a URL with the
scheme "C", which would then fail to load.

In addition to checking whether the resulting URL would be relative
(without scheme), we now also check whether the given string is a valid
absolute file path. In both cases, we produce a URL pointing to the
local file.

Closes #2841