Improved positioning when adding maps to world via context menu
Previously the map would get added to the world at its origin, where it
would likely overlap with existing maps in the world. Now the map is
positioned to the right of the last map that was added to the world,
where it is less likely to overlap.

Also improved the calculation of the map size by using the renderer
instead of only handling the orthogonal map case.