Scripting: Initialize tile layer size to map size upon add
When adding a tile layer to a map, the size of the layer is now
automatically set to the size of the map when it is still on 0x0.

This is mainly for convenience. Not setting the layer's size currently
causes rendering issues and may result in other problems as Normally new
layers automatically match the map's size and there is no way to resize
them, so the scripting API now helps a little to avoid these problems as
well.

Closes #2879