Enabled erasing with the Wang Brush
To select the erasing mode you currently have to capture the empty color
from the map (right-click on an empty area). But once you have that, the
tool allows you to erase.

In erasing mode, the WangFiller consideres the empty tile to be a valid
candidate as well. At the same time, is now sets the "checked" flag on
any cells it sets, to be able to tell them apart from empty cells.