Improved error message when adding external tileset
When no tileset format could be found for the selected file, an empty
error dialog would pop up. Now it says "Unrecognized tileset format".

Also tweaked the behavior when dragging images onto the Tilesets view,
when the New Tileset dialog pops up for the given image, the dialog will
be set to the "Based on Tileset Image" mode. Because it would be really
rare that somebody wants to create an "Collection of Images" with that
one image in it.