Wang Brush: Added mode in which neighbors are adjusted
Now when it turns out a neighboring tiles doesn't fit to a placed one,
and this neighbor is outside of the filled region, it can be queued for
correction. To avoid this growing into a random mess, the transition
penalty is now based on the distance between desired and chosen color of
each index, the same approach used by the Terrain Brush.

At the same time, when corrections are disabled, like when using Wang
mode with the fill tool or Stamp Brush, I've reintroduced the look-ahead
that makes sure that a chosen candidate will not cause an impossible
situation for any of its neighbors. This prevents most of the problems
when filling with incomplete Wang sets.