Introduced the Wang set type
Now a Wang set can be defined to be corner-based, edge-based or both.
This affects the way the Wang Brush works, the overlay drawn by the
Tileset Editor and the process of setting the WangIds on the tiles.

The Wang set type is saved, loaded, editable and generally used, but the
UI should probably be adjusted so that you select the type of Wang set
already when creating it (currently that always creates a "Mixed" set).

Also, the WangIds are not changed when the Wang set type is changed. It
might be a good idea to explicitly clear any WangIds when changing the
type as appropriate, though I'm a bit reluctant to do that due to data
loss.