Wang Brush: Choose a meaningful mode for "Mixed" Wang sets
It should be helpful to avoid trying to make edits that can't work out,
so I've reintroduced some code that determines the brush mode by looking
at where a color is used.

It might still make sense to add an explicit toggle for the user, to
override the "PaintEdgeAndCorner" mode.