Scripting: Fixed reset of file formats on script reload
When the scripts got reloaded, or when a native plugins was disabled &
enabled within the same session, the map, tileset or object template
could lose track of which format they were saved in.

Upon saving a map that lost track of its format, it would silently
default to saving in the TMX format. Now, if the format can't be found
anymore, the user gets a "Save As" dialog instead.

Closes #2911