AppVeyor: Switched the Qt 5.12.9 build from MinGW to MSVC
This build is only done for testing purposes, so let it build with an
alternative compiler to catch potential compilation issues early.

Zstd is not enabled for the Visual Studio build. I'm not sure how to set
that up right now, but it's not important since this build is not
distributed.