Fixed linker error on Visual Studio
The Grid template class should not be marked TILEDSHARED_EXPORT, since
this makes the Visual Studio compiler think that libtiled will provide
the symbols for the instantiatons, but it can't since the instantiations
are not in libtiled.

Actually right now the Grid template could be moved from libtiled to
Tiled, though I hoped to use it in libtiled as well since it closely
matches the way cells are stored in the TileLayer. The main difference
is the way the TileLayer::setCell function has to take into account the
"checked" flag on the cell and the special logic for keeping track of
the used tilesets.