Include moc-generated files where necessary with Qt 6
This fixes compilation against Qt 6, which requires also pointers to
types used as meta type to be fully defined.

When these moc-generated files are compiled on their own the necessary
header files are not included. This is fixed by including them
explicitly in the right context, in which case they will no longer be
compiled on their own.

This fix is preferable over the previously attempted fix for this
problem in commit 1d6df01c00f5d010622bce907365f1d8a5fcf630 (reverted).