GitHub Actions: Added itch.io upload for snapshots
Changed the 7-zip archives to a direct uploading of the install-root
folder. This is the preferred thing to push to itch and it also avoids a
7-zip in a zip download from GitHub.

To ensure a unified version for all snapshots and to make it available
for the upload jobs, the version is now set in a job of its own.

Disabled itch.io upload for the builds on Travis CI and the Windows
64-bit build on AppVeyor. For now, AppVeyor will still upload the
Windows XP build.