GitHub Actions: Ship the necessary OpenSSL libs
These are necessary for update checks and news.

The OpenSSL libraries seem to not be installed for distributing them
along with applications on GitHub Actions (as opposed to AppVeyor). To
have them available I opted to download them from download.qt.io,
extending the install-qt.sh script to be able to install openssl as
well.