GitHub Actions: Removed installation of libqt5svg5-dev
This package is not necessary to compile Tiled, as long as a recent
enough version of Qbs is used since older versions raised an error when
certain referenced files could not be found, even if the references were
from products that were not being built.

Qbs 1.13.1, available on Ubuntu 20.04, is recent enough.