Fixed opening of files in already open instance
When opening files from a file explorer or by launching Tiled from the
command-line with one or more files as arguements, Tiled would only open
those files in an already open instance, when that instance was launched
in similar fashion. Tiled instances launched from a shortcut would never
get reused.

This was due to not calling QtSingleApplication::isRunning, which caused
it to never become a listening instance when launched without files to
open.

Thanks to eishiya on Discord for reporting this issue.