WebAssembly: Two small compile fixes
This makes Tiled compile against Qt for WebAssembly, but it is unusable
because it appears unable to open/save files.

Making this actually work is likely a much larger change. Qt provides
special functions for opening and saving files through a user dialog
(QFileDialog::getOpenFileContent and QFileDialog::saveFileContent), but
those are not a good match with the way Tiled likes to work with files
(for example, opening a file usually requires reading additional
referenced files like tilesets and images).

Apart from file reading problems, there is another issue in that
QSettings is asynchroneous on WebAssembly. Tiled would have to somehow
wait for it to have finished loading, before reading any settings.