Add @mapeditor/tiled-api module for typings and documentation
First draft of a npm package and typedoc based documentation.

The index.d.ts is currently mostly in the state as provided
by Alula on Discord, I only fixed the syntax for some function
declaration, added the minimumal needed API for the example from the
README and added documentation for `registerMapFormat` as a proof of
concept.