Fixed performance issues when using a lot of custom properties
The ResizeToContents resize mode on the QtTreePropertyBrowser resulted
in excessive relayouting while properties were being added/removed.

In addition, the QtTreePropertyBrowser was not set to have a uniform
row height, though it really should have. Setting this to true also
avoids a lot of expensive layout calculations.

Selecting an object with 20 custom properties could take about a second
due to the Properties view update. After this patch it feels instant.

Closes #2958