Added smooth scrolling option
When smooth scrolling is enabled, the arrow keys move the view smoothly
using a fixed speed, as opposed to in steps based on key repeat. This
option is enabled by default for now, though that can be reconsidered
based on the feedback.

At the same time, this change makes it possible to move the view
diagonally by pressing two arrow keys, also when the option is disabled.

Also made tileset view scroll wheel handling support pixelDelta on
platforms that use it, for more precise scrolling (untested).