Renamed "Wang" to "Terrain" in the UI
In the end I've decided, based on helpful feedback, to rename the Wang
related functionality. This keeps the familiar "Terrain" terminology,
which makes it a lot easier to talk about the visible improvements
compared to the previous versions of Tiled.

* "Wang Set" -> "Terrain Set"
* "Wang Color" -> "Terrain"
* "Wang Brush" -> "Terrain Brush"
* "Wang Fill Mode" -> "Terrain Fill Mode"

This change breaks the String Freeze, changing about 20 strings, which
fortunately only affects three translations that have already received
updates.

A few additional changes were made:

* When creating a new Terrain Set the user has to choose the type of the
  set, since this choice otherwise gets lost easily and most users would
  end up with a "Mixed" set which is not the most common.

* New Terrain Sets get one terrain added by default, since an empty set
  is not useful.