YY plugin: Handle tile layers using multiple tilesets and other stuff
Split the export into two steps, where first all the layers are iterated
and converted to GameMaker structures. Serializing stuff to the file
happens as second step. This allows moving the data around as needed
more easily.

As a result of this change, it was now easy to handle a tile layer where
multiple tilesets are used, which will now create a group layer with
a child tile layer for each tileset.

Similarly, this way we can write out asset, instance and path layers for
a single group layer. However, writing of paths and instances is not
fully supported yet.

Also, all 8 views are now always written, as done by GameMaker.