YY plugin: Export tile layers as asset layers when needed
There are a number of features that can't be supported on a GMRTileLayer:

* Image collection tilesets
* Tilesets with a tile size different from the grid size
* Non-orthogonal tile layers

This change falls back to exporting the tiles to an asset layer in such cases instead, adding support for all the above features.

A number of warnings have been added as well.

The "ignore" property for instances is now set to true for invisible objects by default.

Simplified a little the logic of assigning unique names.