YY plugin: Fixed handling of origin in several places
Tile objects with custom origin would not export to the right location when scaled, both when exported as assets and as instances.

Assets created based on tiles from an image collection tileset were not taking the origin into account in determining their position.

There are still some remaining issues, but I'll have to revisit those later.