Avoid setting export.target to "."
The loading code would cause an empty string to be interpreted as the
directory of the map/tileset file, which on saving would get saved as a
"." (relative reference to current directory).

This patch avoids that problem, in addition to fixing up any files which
got a "." written as the export target.