Scripting: Enable extending context menus
Currently five of the context menus are extensible. There are a bunch
more context menus in Tiled, but for most extending them would not make
sense without access to their context. For example, it is currently not
possible to access the clicked tab in case a tab context menu is opened.

The ability to extend context menus is implemented by remembering the
extensions passed to tiled.extendMenu so that they can be later applied
to temporary menus using a call to ActionManager::applyMenuExtensions.