snap: Try alternative fix for getting DBUS to work
The 'dbus' snap plug appears to have a different purpose, and causes
the following warning:

  snap "tiled" has bad plugs or slots: dbus (cannot find attribute 'bus')

Instead, the reason DBUS communication didn't work for the snap was that
the snap is built with qmake and the qmake project wasn't enabling DBUS
yet.

Also removed a reference to the no longer existing 'terrain.h' from the
qmake project.