Fixed Terrain Brush behavior on map edges
The Terrain Tool was not behaving as expected near the edge of the map,
since it ended up trying to modify terrain outside of the map and then
processing "corrections" to neighboring tiles based on locations outside
of the map.

Fixed by limiting the modified area to the map and by avoiding corrections
to be queued for locations outside of the map.

This was a regression introduced in Tiled 1.5.