Fixed Terrain Brush behavior for sets transitioning to nothing
When a terrain transitions to nothing, adjusting neighboring tiles may
require erasing them entirely. However, erasing of tiles was only
enabled when painting with "no terrain". As a result it would place some
arbitrary "almost fitting" tile instead, which was undesirable.

Now erasing is always enabled when editing with the Terrain Brush. It is
still disabled when using the Terrain Fill Mode, where the placement of
empty tiles causes problems.

This was a regression introduced in Tiled 1.5.