Show a helpful text when there are no Terrain Sets
The empty Terrain Sets view now tries to be a little more helpful than
just being empty.

I tried automatically showing/raising the Terrain Sets view when the
Terrain Brush is selected, but it didn't feel right. Could work if it
also automatically goes back to its previous state when the tool is
deselected, but it's not obvious to me how that could be implemented.

Also fixed a crash that could happen when removing a tileset, because
that could delete a still referenced WangColorModel.

Part of issue #3015