Added toggle to select enclosed rather than touched objects
So far the Object Selection Tool always selected any touched object when
using the selection rectangle. This change adds a toggle to the tool bar
that allows changing the behavior such that it only selects the objects
enclosed within the selection rectangle.

The Alt modifier can be used to temporarily switch between these modes.

Icons done by eishiya and Sygmei.

Closes #3023