Render fine grid over main grid with thicker lines
Added gridFine - new argument to drawGrid method in MapRenderer. For drawGrid
this value acts in sense of majorTicks. When gridFine set to non-zero value
Orthogonal and Isometric renderers will draw thichker lines with this step.

HexagonalRenderer does not support this because it's unclear what exactly we
need to render here as the fine grid.