docs: Avoid "namespace" for WangSet enum
Thanks to MrVirite on Discord for figuring out this option, which seems
to be quite a good compromise. It no longer makes the actual values
known, but it makes the generated documentation more readable.

Now there's a lot of enums that should still be added...

Also added a $2 to the generate-scripting-tsdoc.sh, mainly so that one
can pass --watch.