Fixed possible crash due to invalid 'current object' reference
The 'current object' of a Document might be owned by a different
Document instance. This happens when a Tile, WangSet or WangColor
becomes the current object in the MapEditor.

These types are set as current object in order to display their
properties in the Properties view, which is useful even if only in
read-only mode.  However, we do need to make sure to reset the current
object before it gets deleted, otherwise it becomes a roaming pointer
and can later lead to a crash when the Properties view gets updated.

Fixed by not only keeping track of the current object, but when
necessary also of the owning document. This way, relevant events from
the owning document can be handled in order to reset the current object
as needed.

This crash was never directly reported, but about 5 automated crash
reports from the Tiled AppImage seem to be caused by this problem.