Preserve tile order when only some of the tiles are saved
When tiles based on a tileset image are not re-ordered then only the
ones with relevant meta-information are included in the file. Since
these tiles are now added to the mTiles list earlier, they would appear
at the top in the tileset view.

This problem is fixed by inserting each tile at its expected location,
instead of appending it.