Fixed iteration order in AdjustTileMetaData
This reverts part of aac7b4159b26ad58a4f0c29e9933e383b52c3286. In
AdjustTileMetaData we always need to make sure to copy the meta-data
either to larger indices or smaller indices, which is why we can't just
iterate the list of tiles since it can have a custom order.