Custom Commands: Improved starting directory for file chooser
Instead of always starting in the "Applications" directory, which turns
out to be pretty useless on Windows and Linux, be a bit smarter and pick
the current project directory if available, or fall back to the "Home"
directory.

In addition, use the current value or the last picked location when
available.

This change applies to both picking the executable and the working
directory.