Custom Commands: Avoid reset of model when editing command name
The reset clears the selection, which is annoying since we may still
want to edit the command after renaming it.

In addition, when the model is reset there is apparently no
'currentIndexChanged' signal emitted, which left the widgets enabled,
but trying to edit them would cause a crash. An additional safeguard has
been added against this as well.