Cleaned up CustomTypesEditor and made it auto-save values
When adding, removing or changing values, the custom types are now
automatically saved. The explicit "Load" and "Save" buttons have been
removed, as well as the "Debug" button.

Also switched it from using a QTableWidget to a QTreeView and a model
derived from QStringListModel for the values.

Empty values are no longer automatically removed and new values get a
default text.