Moved handling of PropertyValue to PropertyBrowser
This way custom types can be supported much more generally and based on
the existing features in the QtPropertyBrowser solution, requiring less
custom code to support all the various parameters for other types than
enums later on.

One downside is that now custom types no longer work in the Object Types
Editor, because it does not use the PropertyBrowser. Further refactoring
should allow sharing this widget, by taking out the handling of all the
built-in properties.