Make enum properties work in Object Types Editor again
Introduced CustomPropertiesHelper to share some functionality between
the PropertyBrowser and the ObjectTypesEditor, which provide two very
similar editors yet work in a rather different context.

The CustomPropertiesHelper also covers the necessary handling of
PropertyValue / PropertyType.

One remaining issue with the enums is that now when the user changes the
values of the enum, the new values are not reflected in the properties
view. Previously this worked because the QComboBox was re-created, but
now we need to reset the "enumNames" attribute for the created
properties.