Fixed possible crash in Tile Animation Editor
When opening the Tile Animation Editor without having any tile selected,
and then double-clicking a tile (which normally adds it as a frame to the
animation), the application would crash due to referencing a nullptr.

This was a regression in b2761f25c7beedceb4aae, since before that change
there was no tileset visible without having a tile selected.