Scripting: Avoid adding objects with duplicate ID
Reset the object ID when adding an object to a map with an ID that
conflicts with an existing object on that map. This can happen when
using a script to move an object from one map to another, or when a map
file erroneously contained objects with conflicting IDs.

This fix enables a script to be written that automatically resets any
previously duplicated IDs.