Linux: Possible workaround for crash in clipboard manager
Occasionally there's a crash reported on Linux/X11 (through Sentry) when
QXcbClipboard::mimeData is called in response to the
QClipboard::dataChanged signal.

This appears related to QTBUG-22552, but it is closed and was not
reopened based on my comments. Making the connection a queued one might
workaround this issue, but only time will tell whether it will really
stop the crash from happening, since I don't know how to reproduce it
locally.