Fixed data loss when saving maps with tilesets that failed to load
If a tileset fails to load, a dummy tileset is created in its place,
allowing the map to load anyway and any tiles to be rendered with
special markers.

When saving such a map however, the broken tile references were not kept
intact since the "firstgid" of the tileset following such dummy tilesets
would get set to the same value due to the dummy tileset not containing
any tiles.

Fixed by setting the tileset's "nextTileId" based on the tiles which are
actually used, such that when saving the dummy tileset it will get a
range of tile IDs that includes all referenced tiles.

Closes #3106