Fixed copying and capturing stamps on staggered maps
When copying a selected area or creating a tile stamp from the
selection, the captured stamp would only have the tiles in the correct
position when the selection started at even positions.

When capturing an area at an odd position, the stagger index needs to be
adjusted. This was already done for right-click capturing in the
CaptureStampHelper, and now it also happens in
Map::normalizeTileLayerPositionsAndMapSize.

Fixes #2874