Moved object types into the project
When a loaded project referred to an object types file, its types are
automatically imported. This is not done for the global object types
file however, so instructions are probably needed for those who need to
import the file manually (documenting the location of this file).

When the object types file was changed outside of Tiled, it was
automatically reloaded. Automatic reloading still needs to be
implemented for the project.

The new setup is not ideal for those who were generating the object
types file. A potential solution would be to add a scripting API for
registering object types.