Don't require saving maps upon creation
It probably was a mistake to force saving upon creation, since it stands in the
way of quick tests and forces people into organizing their files in a moment
where they may not want to do that.

A similar change is planned for tilesets, where we could instead prevent the
user from using an unsaved tileset in a map.

Eventually, I'd still like to try moving file management largely out of the
user's hands. But for now reverting most of 343e72abd0e4821538bb0 seems
appropriate.

Closes #1902