Apply transformation actions to the preview while placing tiles
Also made sure that the preview object updates when the selected tile
has changed.

One issue with the change as-is, is that it only works while there is a
preview. This means that using the tool bar buttons will currently not affect
the preview since it does not exist when the mouse has exited the map view.

Closes #3091