Tweak the behavior for flipping/rotating while placing tile objects
The flip/rotate will now always apply to the preview, even if it is not
currently active, rather than applying to the selection (usually the
last placed object) in that case.

If you want to actually flip or rotate the selection, you now need to
exit the placement tool (can be done by Escape, for example). I think
this is not a huge loss, because it makes the transformation actions
more consistent and enables clicking the tool bar buttons to activate
them.

Issue #3091