Allow using object context menu in object creation tools
Sometimes one may want to trigger an action on a previously placed
object, and there appears to be no reason to require exiting the
placement tool for that.

The only thing we need to make sure though, is that we're not going to
try selecting any preview or overlay objects. To achieve this I've set
their object group to be locked. Alternatively I could have added an
additional check on the Map*.