Fixed toggling "Clear View" on & off shifting the map
It seems I was activating the wrong layout. By activating the layout of
the MapEditor's widget, the view correctly keeps the same position in
the center. This does mean it may still shift on toggle, but it will now
shift back by the same amount on the next toggle.

This adjusts the logic implemented in fde317e2382a31cc32091b8cdc243276.