Changed text properties to only apply when editing is finished
Previously, when editing any built-in string property (like name and
type) or a custom string or file property, an undo command would be
created for each typed character. While ideally this could be resolved
by merging the undo commands, there are so many of them that I felt this
would not be practical.

Here, I have instead opted to apply the change only when editing is
"finished", which means either the user pressed Enter or the edit field
lost focus.

One thing I'm not happy with, is that in this mode you'd probably expect
the text to not change when pressing Escape, but then it applies as well.

Closes #3103