Fixed serialization of nested custom property types
Rearranged the conversion to and from "property value" to "export
value", such that the original custom class information can be used when
de-serializing the members of that class. This was necessary to
correctly load nested custom classes and enums, because the custom type
is not saved out for member values.

It should also fix cases where the type is lost. In the JSON format, the
information about whether a property value is a path or an object
reference is not saved on member values of custom classes.