YY Plugin: Determine sprite names by looking for meta files
Before assuming the name of an image file without its extension to be
the name of the associated sprite, search the current and up to two
parent directories for a *.yy meta file. If found, use its name
without the file extension instead.
A custom "sprite" property may still override this feature.