Improved multi-layer painting behavior
When painting into multiple layers, and having the same number of tile
layers selected as on the current brush, the layers are now matched based
on their order instead of their name. This has the following advantages:

* It makes sure that capturing or copying part of a map to the same map
  will always end up in the expected location, regardless of the layer
  names.

* It makes it possible to move the content of multiple layers to another
  set of layers, in the common case where the order of these layers stays
  the same.

To make this work by layer count, empty layers are now copied as well in
Map::copyLayers and CaptureStampHelper::endCapture.

When the number of selected tile layers does not match up, the layers are
still matched by name. However, target layers are no longer reused
when the tile stamp contains multiple layers with the same name. In this
case, each consecutive layer with the same name is matched with the next
layer with that name, or one will be created.

When a new layer is created, we now insert it next to the previous or
next matched layer if possible, rather than always adding new layers at
the top.

The logic in StampBrush::drawPreviewLayer and
AbstractTileFillTool::fillWithStamp was adjusted to not rely on searching
by layer name at all, and to create multiple layers with the same name in
the preview, when they existed in the stamp.

Closes #3094