Avoid name collisions in types and class members
Check for existing names when renaming classes and when renaming or
adding members.

Not checking these on enum values for now. While it does not make sense
to have multiple entries with the same name, it also doesn't really
cause any problems. Also, preventing entries with the same name might be
annoying while trying to reorder the values of an enum, for which no
specific actions currently exist.

Part of finishing issue #489