TMX format: Include propertytype attribute for class members
Previously, the custom type information was left out for class members,
because it was essentially redundant. Once we know the type of the
top-level property, the types of the members are known based on the
user-defined class.

However, since the type definitions are stored somewhere else in a
different format (the JSON Tiled project file), it can be very helpful
to include this information anyway.

This currently only affects the TMX format. The JSON and Lua formats
still do not store this information for class members, because it would
make them a lot less convenient to work with.

Closes #3230