Scripting: Fixed possible crash after creating tilesets from script
The EditableTileset instances are sometimes created on-demand by the
EditableManager. However, when creating a new tileset from script by
using "new Tileset(name)", the EditableManager was unaware of this
instance.

This caused the manager to later create another wrapping EditableTileset
for the same tileset, which in turn could cause EditableTile instances
for that tileset to point to different EditableTileset instances. This
could then cause an assertion failure in EditableTileset::detachTiles,
since the EditableTileset instance expects to be the only wrapper for a
given tileset.

Closes #3229