Refactored ImageLayer and ObjectGroup undo commands
The undo commands affecting ImageLayer and ObjectGroup properties are
now based on the "ChangeValue" command as well.

The "ChangeValue" command was improved to be able to change the value of
multiple objects at once, which also resolves the problem where toggling
for example the visibility of a set of layers on and off would fail to
mark the undo command as obsolete.

Part of issue #3103