Avoid deselecting all layers when clicking empty area in Layers view
It's never really useful to deselect all layers, as this only confuses
some of the tools.

Nevertheless, the Eraser tool was tweaked to apply to the "current" tile
layer, even if it is not selected. This situation can now probably only
happen by Ctrl-clicking the currently selected layer, to deselect it.

Closes #2806