Improved the logic for automatically switching tools
Now tools switch automatically when a different layer type is selected,
regardless of whether this causes the current tool to get disabled. And
rather than trying to switch to some previously disabled tool, it now
tries to switch to the previous tool you had selected for the current
layer type.

This might still auto-select the Offset Tool if you switch to a group or
image layer. However, it becomes a lot less likely to be used by
accident, since it also automatically switches back if you select a tile
or object layer again.

Closes #2807