Scripting: Fixed tileset or tile references for maps loaded from script
When a map was loaded with `tiled.mapFormat("tmx").read(...)`, it was
not possible to reliably get tileset or tile references from it, for
example when calling `TileMap.usedTilesets` or `TileLayer.tileAt`. This
only worked when the referenced tileset was already in use by another
open map or was open by itself.

Fixed by always allowing the creation of a wrapping EditableTileset
instance, for which ownership is transferred when eventually a
TilesetDocument is created that should own it.