Increased minimum supported Qt version to 5.9
Previously, the minimum supported Qt version was 5.6. After this change,
Windows XP can no longer be supported, because Qt 5.6 was the last Qt
version to support Windows XP. The respective build on AppVeyor has been
removed.

The minimum supported version is Qt 5.9 now, because the snap package is
still built with this version. The plan is to increase the minimum version
further once the snap package has been updated to a newer version.