Scripting: Automatically add tilesets to the map where needed
It should no longer be necessary to manually add tilesets to maps when
using their tiles on tile layers or as the tile of a tile object. They
will now always be added automatically.

This fixes repainting issues for the preview brush, which did not
calculate its bounding box correctly when the map was unaware of the
tilesets used.

It may also fixed memory issues. We need to make sure that any tileset
used by a map is kept alive using a strong reference.

Closes #3268