Fixed possible crash in tile stamp preview
A crash reported on Sentry (TILED-39) indicated that likely the
'currentLayer' was nullptr when this code executed. Since there is no
hard guarantee that there is always a current layer (and possibly not
even that mMapDocument is always set), I've added some additional
checks.

Potentially the code needs updating in respect to editing multiple
layers, but for now this should hopefully avoid the crash.