Fixed custom types not being usable without opening a project
When custom types were added before opening a project, they would not be
available in the Add Property dialog.

Now they are immediately usable, but care still needs to be taken that
the types are not automatically saved when used without having a project
loaded. Likely future versions of Tiled will not allow adding types
before creating a project.

Closes #3295