Qbs: Enable building against Zstandard using pkg-config
It is disabled when setting project.enableZstd, which continuous to
perform a static linkage against an expected zstd compile in a "zstd"
directory in the root of the repository (used for some builds).

Replaced the define indicating whether Zstandard support is enabled with
a runtime check.

This enables supporting Zstandard in the snap when built with Qbs.

Issue #1484