Qbs/snap: Fix linking issue with zstd
Also removed likely no longer needed -Wno-unknown-pragmas on macOS, to
simplify things a bit.

Renamed the enableZstd option to staticZstd, to avoid some confusion.
Zstd is now enabled by default when detected through pkg-config, and the
staticZstd option is meant to be used for statically linking the library
on macOS and Windows. It can now also be used to enable static linking
of zstd through pkg-config.