Introduced libtilededitor
This change moves most of Tiled into a library, mostly to enable
autotests to link with it, for testing purposes.

Eventually, it might also enable the use of Tiled editor widgets in
another application, or sharing of editor-specific functionality with a
Qt Quick based version of Tiled.

To avoid complicating git history and merging, the source code largely
remains at the same location, with "src/tiled" containing the new
library and "src/tiledapp" now containing the application.