Scripting: Added -e,--evaluate to run a script from command-line
As for extensions, .js files will be evaluated regularly, whereas .mjs
files will be loaded as a module (and can use import/export).

In this case, the Tiled UI will not be instantiated. This means
functions that rely on the UI being present will not work (like
`tiled.open`, `tiled.mapEditor`, etc.). However, scripts should be able
to load, modify and save maps and tilesets through the available formats
(`tiled.mapFormats` and `tiled.tilesetFormats`).

Non-option arguments can be passed to the script by passing them after
the script file name, and are accessible through
`tiled.scriptArguments`.