docs: Fixed custom tool example
`new MapObject(++this.counter)` used to create an object with the number
as the name, but at some point another constructor was added, and this
number is now interpreted as the object type.

We could explicitly convert the number to a string in JS, but I've opted
for calling the constructor that sets the object type to Rectangle.