AutoMapping: Applying rules without "inputnot" layers is now much faster
This optimization was done by lazily computing the list of tiles used in
a rule's input layers, which avoids doing needless computations and
memory allocations because most of the time, this list is not actually
used.

The AutoMapping is 2-3x faster in my benchmarks, but the exact speedup
will depend heavily on the rules. No speedup is expected for rules that
use "inputnot" layers, because they already didn't need to compute this
list.

Also dropped an early-out check that could never happen. When an
InputConditions instance is created, it always has either an input or
an inputnot layer.