AutoMapping: Made region layers optional
When no input or output regions defined, the respective regions will
now be derived from the "input/inputnot" and "output" layers directly.
For many rules, this means the definition of these layers can already be
skipped.

This change does not affect the behavior in case of empty tiles within
the input region, but as part of issue #3100, such tiles will likely be
entirely ignored. All current AutoMapping autotests passed with this
change also with their region layers removed, but they do not currently
test this behavior.

Closes #1918