AutoMapping: Optimized calculation of each rule's input/output region
A quadratic amount of intersection checks were done here. Turns out it
is faster to just perform the two needed intersections for each rule.

This can make a big difference in rule maps with lots of rules (making
the entire loading of rules twice as fast in my test case).