AutoMapping: Made the rule matching a lot faster
The matching was entirely rewritten based on a new data structure. The
"RuleInputSet" struct provides a per-rule tightly packed definition of
its input, which reduces cache misses that are otherwise caused by
looking up cells from various locations in the rule map.

In addition, the pre-processing of rules means we can discard empty cells
or entire rules early on, rather than every time we're trying to match
a rule.

Finally, a small optimization was achieved by using a function pointer
for the wrapping and overflowing options, instead of checking these options
each time we need to get a tile from the target map.

Warning: There's a regression hiding in this change still...